requirements.txt
streamlit
python-docx
beautifulsoup4
lxml
pandas
openpyxl
pyproj
num2words
